:recycle: useSearch 훅으로 분리
